# src/main/resources/api/openapi.yaml
openapi: 3.0.1
info:
  title: TearsDeepMind API
  description: Automated crawling service for financial forums analysis.
  contact:
    name: TearsDeepMind Team
    email: support@tearsdeepmind.com
  license:
    name: Apache 2.0
    url: http://springdoc.org
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
tags:
- name: Crawler Operations
  description: Endpoints for triggering and monitoring forum crawling tasks.
paths:
  /api/v1/crawler/extract/{seccion}/{dias}:
    get:
      tags:
      - Crawler Operations
      summary: Extract threads from a specific section
      description: Triggers an asynchronous crawling task to extract forum threads
        from the specified section for the given number of days.
      operationId: extract
      parameters:
      - name: seccion
        in: path
        description: "The section name to crawl (e.g., DailyAnalysis, QuantUpdates)"
        required: true
        schema:
          type: string
        example: DailyAnalysis
      - name: dias
        in: path
        description: Number of days to look back for threads
        required: true
        schema:
          type: integer
          format: int32
        example: 1
      responses:
        "200":
          description: Extraction task started successfully.
        "500":
          description: Internal server error starting the task.
  /api/v1/crawler/check/{seccion}:
    get:
      tags:
      - Crawler Operations
      summary: Check for new content
      description: Scans the specified section feed and returns a list of thread titles
        that have not been downloaded to the local filesystem for the current day.
      operationId: check
      parameters:
      - name: seccion
        in: path
        description: "The section to check (e.g., DailyAnalysis, QuantUpdates)"
        required: true
        schema:
          type: string
        example: DailyAnalysis
      responses:
        "500":
          description: Internal server error during check.
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: string
        "200":
          description: Check completed successfully. Returns list of new thread titles.
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: string
components: {}